use a base class
