+++ web/modules/weather_blocks/src/Plugin/Block/HourlyForecastBlock.php.test
@@ -2,12 +2,9 @@
 
 namespace Drupal\weather_blocks\Plugin\Block;
 
-include_once "HourlyForecastBlock.php";
-
 use Drupal\Core\Routing\RouteMatchInterface;
 use Drupal\weather_data\Service\WeatherDataService;
 use PHPUnit\Framework\TestCase;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Tests for the HourlyForecast block.
@@ -55,39 +52,6 @@ final class HourlyForecastBlockTest extends TestCase {
 
     $this->hourlyForecastBlock = new HourlyForecastBlock([], '', $definition, $this->weatherData, $this->routeMock);
 
-  }
-
-  /**
-   * Test that the create() static method gives us back a block object.
-   */
-  public function testCreation() : void {
-    $container = $this->createStub(ContainerInterface::class);
-    $container->method('get')->will($this->returnCallback(function () {
-      $args = func_get_args();
-      switch ($args[0]) {
-        case 'weather_data':
-          return $this->weatherData;
-
-        case 'current_route_match':
-          return $this->routeMock;
-
-        default:
-          return NULL;
-      }
-    }));
-
-    $actual = HourlyForecastBlock::create($container, [], '', ["provider" => "weather_blocks"]);
-    $this->assertEquals(HourlyForecastBlock::class, $actual::class);
-  }
-
-  /**
-   * Test that block cacheing is disabled.
-   */
-  public function testCache() : void {
-    $expected = 0;
-    $actual = $this->hourlyForecastBlock->getCacheMaxAge();
-
-    $this->assertEquals($expected, $actual);
   }
 
   /**
