+++ web/modules/weather_blocks/src/Plugin/Block/WeatherBlockBase.php.test
@@ -5,13 +5,17 @@ namespace Drupal\weather_blocks\Plugin\Block;
 use Drupal\Core\Routing\RouteMatchInterface;
 use Drupal\weather_data\Service\WeatherDataService;
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
- * Tests for the CurrentConditions block.
+ * Tests for the WeatherBlockBaseTest block.
  */
-final class CurrentConditionsBlockTest extends TestCase {
+final class WeatherBlockBaseTest extends TestCase {
   /**
    * A ready-to-use current conditions block object.
+   *
+   * We use this block for our tests because the WeatherBlockBase class is
+   * abstract and cannot be directly instantiated. So, we'll test by proxy.
    *
    * @var currentConditionsBlock
    */
@@ -55,38 +59,36 @@ final class CurrentConditionsBlockTest extends TestCase {
   }
 
   /**
-   * Test that the block returns the expected data if we're on a grid route.
+   * Test that the create() static method gives us back a block object.
    */
-  public function testBuild() : void {
-    $this->weatherData->method('getCurrentConditions')->willReturn('this is weather data');
+  public function testCreation() : void {
+    $container = $this->createStub(ContainerInterface::class);
+    $container->method('get')->will($this->returnCallback(function () {
+      $args = func_get_args();
+      switch ($args[0]) {
+        case 'weather_data':
+          return $this->weatherData;
 
-    $expected = [
-      "#theme" => "weather_blocks_current_conditions",
-      "#data" => "this is weather data",
-    ];
-    $actual = $this->currentConditionsBlock->build();
+        case 'current_route_match':
+          return $this->routeMock;
 
-    $this->assertEquals($expected, $actual);
+        default:
+          return NULL;
+      }
+    }));
+
+    $actual = CurrentConditionsBlock::create($container, [], '', ["provider" => "weather_blocks"]);
+    $this->assertEquals(CurrentConditionsBlock::class, $actual::class);
   }
 
   /**
-   * Test that the block returns null if we're not on a grid route.
+   * Test that block cacheing is disabled.
    */
-  public function testBuildNotGridRoute(): void {
-    $definition = [
-      "provider" => "weather_blocks",
-    ];
+  public function testCache() : void {
+    $expected = 0;
+    $actual = $this->currentConditionsBlock->getCacheMaxAge();
 
-    $this->weatherData = $this->createStub(WeatherDataService::class);
-
-    $this->routeMock = $this->createStub(RouteMatchInterface::class);
-    $this->routeMock->method('getRouteName')->willReturn("weather_routes.not-grid");
-
-    $this->currentConditionsBlock = new CurrentConditionsBlock([], '', $definition, $this->weatherData, $this->routeMock);
-
-    $actual = $this->currentConditionsBlock->build();
-
-    $this->assertEquals(NULL, $actual);
+    $this->assertEquals($expected, $actual);
   }
 
 }
